fix: restore dynamic asset admin page
